Version number changed to 1.0.0 - beta, Added ability to make shortcuts appear as images
After picking the file/app to launch it will ask for a png image. It will then copy and resize that image to the main save directory. Then it will load it instead of the red cube if the icon_path value of the selected shortcut instance is not null. Currently you are forced to select an image, but I will make that optional.