Fixed more bugs in the error handling when loading profiles.
Made it default to "null" if icon_path cannot be found instead of just ending the load attempt. Also, in create_loaded_shortcuts(), it was checking if ipath was "null" and if the ipath WAS a file. It should have been checking if it was NOT a file. based on that it would never create the default red square because ipath can't be null and point to a file at the same time.